[#3319] fix(PyClient): ./gradlew test -PskipITs failed in python-client (#3322)
### What changes were proposed in this pull request?

1. Add `skipITs` build params in the Python Client Gradle script.
2. Use different directory storage integration test and unit tests.

### Why are the changes needed?

When running this command ./gradlew test -PskipITs directly. the Gradle
python test task will be failed because of lacking compileDistribution
step. we should fix it to make all the commands work.

Fix: #3319

### Does this PR introduce _any_ user-facing change?

N/A

### How was this patch tested?

CI Passed
