+++ clients/client-python/build.gradle.kts
@@ -51,25 +51,38 @@ tasks {
     args = listOf("./gravitino", "./tests")
   }
 
+  val integrationTest by registering(VenvTask::class) {
+    doFirst {
+      gravitinoServer("start")
+    }
+
+    venvExec = "python"
+    args = listOf("-m", "unittest")
+    workingDir = projectDir.resolve("./tests/integration")
+    environment = mapOf(
+      "PROJECT_VERSION" to project.version,
+      "GRAVITINO_HOME" to project.rootDir.path + "/distribution/package",
+      "START_EXTERNAL_GRAVITINO" to "true"
+    )
+
+    doLast {
+      gravitinoServer("stop")
+    }
+  }
+
+  val unitTests by registering(VenvTask::class) {
+    venvExec = "python"
+    args = listOf("-m", "unittest")
+    workingDir = projectDir.resolve("./tests/unittests")
+  }
+
   val test by registering(VenvTask::class) {
+    dependsOn(pipInstall, pylint, unitTests)
+
     val skipPyClientITs = project.hasProperty("skipPyClientITs")
-    if (!skipPyClientITs) {
-      doFirst {
-        gravitinoServer("start")
-      }
-
-      venvExec = "python"
-      args = listOf("-m", "unittest")
-      workingDir = projectDir.resolve(".")
-      environment = mapOf(
-        "PROJECT_VERSION" to project.version,
-        "GRAVITINO_HOME" to project.rootDir.path + "/distribution/package",
-        "START_EXTERNAL_GRAVITINO" to "true"
-      )
-
-      doLast {
-        gravitinoServer("stop")
-      }
+    val skipITs = project.hasProperty("skipITs")
+    if (!skipITs && !skipPyClientITs) {
+      dependsOn(integrationTest)
     }
   }
